hive-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

httpx==0.27.2

faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
numpy==1.26.4

pypdf==5.0.1
python-docx==1.1.2

# RAG enhancements
rank-bm25==0.2.2

# Voice processing
openai-whisper
pydub
python-multipart
openai>=1.0.0
